add validation whether entity exists

+++ src/main/java/org/bookstore/service/impl/BookServiceImpl.java
@@ -45,6 +45,9 @@ public class BookServiceImpl implements BookService {
 
     @Override
     public void delete(Long id) {
+        if (!bookRepository.existsById(id)) {
+            throw new EntityNotFoundException("Can't find book with id: " + id);
+        }
         bookRepository.deleteById(id);
     }
